rename minio_gateway to minio_repo #5
